requirements.txt
pandas
matplotlib
mplfinance
yfinance
openai
python-dotenv
PyYAML
logging
